added MapGeneration and display of harbors

+++ src/main/java/com/aviumauctores/pioneers/controller/GameReadyController.java
@@ -373,6 +373,8 @@ public class GameReadyController extends PlayerListController {
                 Color.CHOCOLATE);
         updateComboBox();
 
+        gameService.setUpdateOption(2, 10);
+
         return parent;
     }
 
@@ -475,13 +477,15 @@ public class GameReadyController extends PlayerListController {
             memberAlert.setTitle(bundle.getString("warning"));
             memberAlert.setHeaderText(null);
             Optional<ButtonType> result = memberAlert.showAndWait();
-            if (result.get() == proceedButton) {
-                gameMemberService.deleteMember(userService.getCurrentUserID())
-                        .observeOn(FX_SCHEDULER)
-                        .subscribe();
-            } else {
-                memberAlert.close();
-                return;
+            if (result.isPresent()) {
+                if (result.get() == proceedButton) {
+                    gameMemberService.deleteMember(userService.getCurrentUserID())
+                            .observeOn(FX_SCHEDULER)
+                            .subscribe();
+                } else {
+                    memberAlert.close();
+                    return;
+                }
             }
         }
         gameService.setCurrentGameID(null);
@@ -625,7 +629,7 @@ public class GameReadyController extends PlayerListController {
 
     public void setMapsizeAndVictorypoints(int victoryPoints, int mapSize) {
         if (mapSize >= 0 && mapSize <= 10 && victoryPoints >= 3 && victoryPoints <= 15) {
-            gameService.setUpdateOption(victoryPoints, mapSize).subscribe();
+            gameService.setUpdateOption(mapSize, victoryPoints).subscribe();
         } else {
             ButtonType proceedButton1 = new ButtonType("OK", ButtonBar.ButtonData.OK_DONE);
             Alert alert = new Alert(Alert.AlertType.CONFIRMATION, bundle.getString("failedEntering"), proceedButton1);
